Improved Lava Charm and similar Accessories. Improved Obsidian Rose and similar Accessories.
Reduce lava damage no longer overwrites the whole method and should now be a lot more compatible with other mods and Minecraft itself.

Lava immunity is now done in a mixin, it also has a component. This allows for lava immunity without fire resistance (so they can work together), and allows for the immunity to refill like in Terraria. Overall works a lot better and is more accurate to Terraria.