Update to 1.3.3.
Fixed missing chest on Floating Islands.
Fixed crash with Bumblezone and possibly other mods (changed how Demon Eye spawns)

Some cleanup here and there, changed uuid of some attributes and moved to be in same class instead of TerraMine.class.

Added Ironskin potion and added it in addition to the Mining potion to the correct chest loot tables.

Started working on global values stored in saves for storing if in hardmode and which ores should generate per world.