1.19 port.
Seems to work from my small amount of testing, had to remove shield for now as Shield Lib isn't updated yet but will implement it in another way anyway because that library is jank.

Villager professions are not working fully yet either, for some reason the villages will not get the profession when near the tinker table but the profession does exist and works when summoned using command and villager with profession will notice the tinker table and stay near it. very odd.