Added Dungeon. The top and stairs now generate, still more work to be done such as processor for random cracked and fancy block placement. Need to figure out how to get rooms to generate at the bottom of the stairs as well and then create the other two color variants.
Dungeon Bricks are now indestructible when below y 60 unless using Netherite Pickaxe or greater.
This is still wip, can be used to troll people at the moment as it is indestructible even when placed by a player (which is not a good).

Now using Fabrics mining level tag system for blocks, makes more compatible with other mods and even this mod once I have terraria tools.