Blue and Green variant of dungeon. Also got created the processor for random cracked and fancy bricks.
Still need to get rooms generating at bottom of stairs, I have one idea of how to do it, but I don't know for certain if it will work, and it is a bit inefficient. Might have to do some more research or ask for help somewhere, but I'll see what I can do.