New Advancements. Dungeon Crawler and Dead Men Tell No Tales, one being from Terraria and the other not. Also updated the Corruption advancement to include the desert variant.
Instant TNT can now be summoned (was trying to get the tnt killing the player to grant the advancement and this was a by product of this, not like that worked anyway as it just uses opening the trapped chest now).

Hardmode state can be set and retrieved with a command (not currently used but good for testing when I get to that stuff).

Going to try to improve the Rocket Boot system tomorrow and use that same system for wings, shouldn't take too long to do, hopefully.