Added raw demonite block (uses the texture for the demonite block, that now has a different texture).
Added craft to turn blocks back into their craft components (ingots and raw).

Demonite can now also be used to activate beacons (doesn't show that it can, but that's because Mojang hard coded the menu, shows that it can in rei though).

Thinking of adding more entities, having only the Demon Eyes is a bit sad, so I will try to add a bunch of entities for the next major update (maybe even a boss?).