Added Devourer (Heavy WIP, needs a lot of work). ReAdded Corruption (and now Crimson) Spread.
Evil biome spread now works correctly, it actually sets the biome so unlike the old broken method the spread to biome will now spawn the correct mobs and for mods such as Travellers Titles should now display the right biome.
May need to optimise it further somehow (as well as the actual block spread) but corruption spread is going to be laggy in Minecraft no matter what because of it being 3D, this is why it can be disabled.