Dynamite works perfectly now. Added Sticky and Bouncy Dynamite.
Actually really fun throwing dynamite, would highly recommend throwing a bunch of bouncy dynamite.

The base is also ready for Grenades and Bombs to be easily made, I'll get the models and textures done for them probably tomorrow and then quickly create the classes needed for them to work.

After this, I want to get the throwing knifes and spiky balls done, which shouldn't be too difficult now that I know how to make throwable projectiles.

Last, I'll try to get the Meteorites done, they will randomly fall on loaded chunks and will create an explosion that generates meteorite ore, this may take a while to figure out, so I may save it for the next update.

I also want to try backport this update to 1.18 since a lot of people still use that version, won't do any earlier versions until the mod is closer to being done.