All meteorite blocks and items added. Did some cleanup with the recipes and textures, should be much easier to find things now.
Just need to get the meteorite to randomly fall onto the world, that is going to be some work to get it to spawn even in unloaded chunks (unless I can't do that, in which case then just loaded chunks) and not crash into people houses.

Terraria uses certain blocks to figure out what area must be a players house, so I'll do something similar but maybe also add a meteorite shield block that also prevent the meteorite from falling.