Preparing for 1.7.0. Added a whole new GUI for easy access to accessories, vanity, and dyes. GUI will also be used for various teams stuff later (teams are going to be very important for servers). Also plan to update Minecraft to 1.19.4, got some atlas stuff in place but will probably need to do a lot for the update.
I plan to remove Trinkets as a dependency, the mod is cool and easy to use, but I just don't have enough access to the slots for what I need. It's going to take a while, but I'm going to try to add my own system to make the accessories function.