Fixed GUI! Now you can put stuff in the slots correctly, the GUI still has some issues like clicks sometimes not registering or for some reason if you hold down the crouch key for a few seconds you will crouch even though you have a GUI open and that should happen. GUI is also now accessible from Creative Mode (still wip, some slight jank right now)
Working on the ability to add more slots during gameplay for the Demon Heart that will be dropped from the Wall of Flesh, Hardcore mode will also add an extra slot by default but that doesn't need to be done during gameplay.

Then I will add the rest of the slots to the GUI and start on adding functionality to the slots.

I really want to get this GUI done because I want to get started on the boss system (which I have fully planned out and is going to work incredibly well on big servers, no trolling other players by summoning moon lord on them!).
Also got to finish some other stuff that I had started before my break, like the meteorite spawning system.