Added teams system in Accessory menu. Added armor value to Accessory menu. Fixed hover highlighting on accessory visibility buttons (player rotate arrows still have this issue because mojang not big brain, I'll make my own widget to fix it).
Currently, teams don't do anything, but the system is there.
I do need to add support for FTB Teams once that's updated for 1.20.1 (will use that system instead of my one).
Next is getting the test boss to only target, be able to attack, and only be attacked by people on the same team.
Then I can create a test spawn item and make it so the team the boss checks is that of the one that spawned the boss.
Also, probably add a check on boss spawn to find all players on the same team within a certain range.