Improved Treasure Bag loot filling.
The Eye of Cthulhu's treasure bag is the only one that drops different items depending on the world's evil type in Terraria, so I added that functionality.
I also made it so that all the items of the same type are stacked instead of something like demonite filling the whole bag.
And last but not least, I changed how the loot table filled the bag, now it selects every option only once and will use the weight to determine if the item should be added or not. This means I can guarantee items are added to the bag and make sure the player doesn't get more of an item then they should.

I would like if these where already a thing you could do with loot tables using just the json but ah well.